In the JNI docs, note that the sizeof SQLITE_TRANSIENT and SQLITE_STATIC should ideally be the same as the platform's pointer size.
FossilOrigin-Name: 7d4ac44ec419ed0474bdb9d237b97660cf0d8faba8fe686f6a914d7bc04dfa3b

+++ ext/jni/src/org/sqlite/jni/SQLite3Jni.java
@@ -849,9 +849,11 @@ public final class SQLite3Jni {
   public static final int SQLITE_SAVEPOINT = 32;
   public static final int SQLITE_RECURSIVE = 33;
 
-  // blob finalizers:
-  public static final int SQLITE_STATIC = 0;
-  public static final int SQLITE_TRANSIENT = -1;
+  // blob finalizers: these should, because they are treated as
+  // special pointer values in C, ideally have the same sizeof() as
+  // the platform's (void*), but we can't know that size from here.
+  public static final long SQLITE_STATIC = 0;
+  public static final long SQLITE_TRANSIENT = -1;
 
   // changeset
   public static final int SQLITE_CHANGESETSTART_INVERT = 2;
